# eliminate clinchedDBValues class