--- a/siteupdate/cplusplus/functions/sql_file.h
+++ b/siteupdate/cplusplus/functions/sql_file.h
@@ -1,11 +1,5 @@
-class ClinchedDBValues;
 class ElapsedTime;
-class GraphListEntry;
-class HighwaySystem;
-class Region;
-class TravelerList;
 #include <array>
-#include <iostream>
 #include <list>
 #include <mutex>
 #include <vector>
@@ -14,7 +8,6 @@ void sqlfile1
 (	ElapsedTime*,
 	std::vector<std::pair<std::string,std::string>>*,
 	std::vector<std::pair<std::string,std::string>>*,
-	ClinchedDBValues*,
 	std::list<std::string*>*,
 	std::list<std::string*>*,
 	std::mutex*