@yakra 2022-04-01
Let's wait a tick on merging this in.
I've come across some more efficiency improvements I can make to the userlog routines.
Getting rid of `csmbr_values` is a prerequisite. If this works out well, it'll make including the Python changes a foregone conclusion. :)